+++ src/lib/db/models/mitmAlias.ts
@@ -1,0 +1,32 @@
+/** db/models/mitmAlias.ts — MITM alias CRUD (mitmAlias namespace). */
+
+import { getDbInstance } from "../core";
+import { backupDbFile } from "../backup";
+import { getKeyValue } from "./shared";
+
+export async function getMitmAlias(toolName?: string) {
+  const db = getDbInstance();
+  if (toolName) {
+    const row = db
+      .prepare("SELECT value FROM key_value WHERE namespace = 'mitmAlias' AND key = ?")
+      .get(toolName);
+    const value = getKeyValue(row).value;
+    return value ? JSON.parse(value) : {};
+  }
+  const rows = db.prepare("SELECT key, value FROM key_value WHERE namespace = 'mitmAlias'").all();
+  const result: Record<string, unknown> = {};
+  for (const row of rows) {
+    const { key, value } = getKeyValue(row);
+    if (!key || value === null) continue;
+    result[key] = JSON.parse(value);
+  }
+  return result;
+}
+
+export async function setMitmAliasAll(toolName: string, mappings: unknown) {
+  const db = getDbInstance();
+  db.prepare(
+    "INSERT OR REPLACE INTO key_value (namespace, key, value) VALUES ('mitmAlias', ?, ?)"
+  ).run(toolName, JSON.stringify(mappings || {}));
+  backupDbFile("pre-write");
+}
