+++ src/lib/db/models/shared.ts
@@ -1,0 +1,19 @@
+/** db/models/shared.ts — foundational JSON-record helpers shared across model sub-modules. */
+
+export type JsonRecord = Record<string, unknown>;
+
+export function asRecord(value: unknown): JsonRecord {
+  return value && typeof value === "object" && !Array.isArray(value) ? (value as JsonRecord) : {};
+}
+
+export function toNonEmptyString(value: unknown): string | null {
+  return typeof value === "string" && value.trim().length > 0 ? value.trim() : null;
+}
+
+export function getKeyValue(row: unknown): { key: string | null; value: string | null } {
+  const record = asRecord(row);
+  return {
+    key: typeof record.key === "string" ? record.key : null,
+    value: typeof record.value === "string" ? record.value : null,
+  };
+}
